# internal/httpapi/server.go
// Package httpapi is the user-facing REST + SSE surface consumed by the SPA.
package httpapi

import (
	"encoding/json"
	"log/slog"
	"net/http"
	"reflect"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"

	"github.com/benlik386/pinkglasses/internal/audit"
	"github.com/benlik386/pinkglasses/internal/auth"
	"github.com/benlik386/pinkglasses/internal/config"
	"github.com/benlik386/pinkglasses/internal/domain"
	"github.com/benlik386/pinkglasses/internal/launch"
	"github.com/benlik386/pinkglasses/internal/obj"
	"github.com/benlik386/pinkglasses/internal/planner"
	"github.com/benlik386/pinkglasses/internal/store"
)

// Server holds dependencies for the API handlers.
type Server struct {
	st       *store.Store
	planner  *planner.Planner
	audit    *audit.Logger
	hub      *SSEHub
	obj      *obj.Store // wordlist uploads land here, not in the database
	logins   *loginLimiter
	launcher *launch.Launcher
	routes   []RouteInfo // filled by Routes(), see RouteTable
}

// New builds the API server.
func New(st *store.Store) *Server {
	s := &Server{
		st:      st,
		planner: planner.New(st),
		audit:   audit.New(st),
		hub:     NewSSEHub(),
		obj:     obj.New(config.LoadAPI().S3),
		logins:  newLoginLimiter(),
	}
	s.launcher = launch.New(st, s.planner)
	return s
}

// Routes returns the HTTP handler for the api binary.
func (s *Server) Routes() http.Handler {
	s.routes = nil
	r := chi.NewRouter()
	r.Use(middleware.RequestID)
	r.Use(middleware.RealIP)
	r.Use(middleware.Recoverer)
	r.Use(securityHeaders)

	r.Get("/healthz", func(w http.ResponseWriter, _ *http.Request) { w.Write([]byte("ok")) })

	r.Route("/api/v1", func(r chi.Router) {
		// The only unauthenticated endpoints. `status` says whether the install
		// needs its first administrator; `setup` creates that one account and
		// refuses once any account exists; `login` is the front door.
		pub := roled(r, RolePublic, s)
		pub.Get("/auth/status", s.authStatus)
		pub.Post("/auth/setup", s.setup)
		pub.Post("/auth/login", s.login)

		// Everything past here needs an identity. The whole surface is wrapped
		// at once rather than endpoint by endpoint, so an endpoint somebody
		// forgets to guard is unreachable rather than public.
		r.Group(func(r chi.Router) {
			r.Use(s.requireAuth)

			any := roled(r, RoleAnySignedIn, s)
			any.Post("/auth/logout", s.logout)
			any.Get("/auth/me", s.me)
			any.Post("/auth/password", s.changePassword)

			// --- viewer: read everything, change nothing ---
			r.Group(func(r chi.Router) {
				r.Use(require(auth.RoleViewer))
				v := roled(r, "viewer", s)

				v.Get("/scopes", s.listScopes)
				v.Get("/scopes/{scopeID}/summary", s.scopeSummary)
				v.Get("/scopes/{scopeID}/targets", s.listTargets)
				v.Get("/scopes/{scopeID}/target-groups", s.listTargetGroups)

				v.Get("/scan-params", s.listScanParamSpecs)
				v.Get("/scopes/{scopeID}/scan-profiles", s.listScanProfiles)
				v.Get("/scopes/{scopeID}/runs", s.listRuns)
				v.Get("/scopes/{scopeID}/schedules", s.listSchedules)
				v.Get("/runs/{runID}", s.getRun)
				v.Get("/runs/{runID}/targets", s.runTargets)
				v.Get("/runs/{runID}/events", s.runEvents)
				v.Get("/runs/{runID}/activity", s.runActivity)
				v.Get("/runs/{runID}/diff", s.runDiff)
				v.Get("/runs/{runID}/footprint", s.runFootprint)

				v.Get("/scopes/{scopeID}/domains", s.listDomains)
				v.Get("/scopes/{scopeID}/graph", s.domainGraph)
				v.Get("/scopes/{scopeID}/hosts", s.listHosts)
				v.Get("/scopes/{scopeID}/hostrows", s.listHostRows)
				v.Get("/hosts/{ipID}", s.hostDetail)
				v.Get("/hosts/{ipID}/services", s.hostServices)
				v.Get("/services/{serviceID}/screenshot", s.serviceScreenshot)
				v.Get("/scopes/{scopeID}/search", s.search)
				v.Get("/scopes/{scopeID}/search/facets", s.searchFacets)
				v.Get("/search", s.searchGlobal)
				v.Get("/search/facets", s.searchFacets)
				v.Get("/scopes/{scopeID}/findings", s.listFindings)

				v.Get("/scopes/{scopeID}/notifications", s.listChannels)
				v.Get("/scopes/{scopeID}/notifications/deliveries", s.listDeliveries)
				v.Get("/wordlists", s.listWordlists)
				v.Get("/wordlists/{wordlistID}/content", s.getWordlistContent)
				v.Get("/workers", s.listWorkers)
				// Where a run may scan from: every pool that is not a run's own.
				v.Get("/pools", s.listPools)

				// A tunnel's name and last egress, never its body. Reading
				// which exits exist is not the same as being able to use one.
				v.Get("/scopes/{scopeID}/vpn-configs", s.listVPNConfigs)

				// Your own tokens; an administrator sees everyone's.
				v.Get("/tokens", s.listTokens)
				v.Post("/tokens", s.createToken)
				v.Delete("/tokens/{tokenID}", s.revokeToken)
			})

			// --- operator: everything about finding things ---
			r.Group(func(r chi.Router) {
				r.Use(require(auth.RoleOperator))
				o := roled(r, "operator", s)

				o.Post("/scopes", s.createScope)
				o.Post("/scopes/{scopeID}/targets", s.addTarget)
				o.Post("/scopes/{scopeID}/target-groups", s.createTargetGroup)
				o.Patch("/scopes/{scopeID}/target-groups/{groupID}", s.patchTargetGroup)
				o.Delete("/scopes/{scopeID}/target-groups/{groupID}", s.deleteTargetGroup)
				o.Patch("/scopes/{scopeID}/targets/{targetID}", s.patchTarget)
				o.Delete("/scopes/{scopeID}/targets/{targetID}", s.deleteTarget)
				o.Post("/scopes/{scopeID}/scan-profiles", s.saveScanProfile)
				// Starting a run sends packets at somebody's infrastructure,
				// which is why it is the boundary between reading and acting.
				o.Post("/scopes/{scopeID}/runs", s.createRun)
				o.Post("/runs/{runID}/cancel", s.cancelRun)
				o.Post("/runs/{runID}/pause", s.pauseRun)
				o.Post("/runs/{runID}/resume", s.resumeRun)
				o.Post("/runs/{runID}/rerun", s.rerunRun)
				o.Delete("/runs/{runID}", s.deleteRun)
				// Recurring scans, and the company default exit they use.
				o.Post("/scopes/{scopeID}/schedules", s.createSchedule)
				o.Patch("/schedules/{scheduleID}", s.patchSchedule)
				o.Delete("/schedules/{scheduleID}", s.deleteSchedule)
				o.Patch("/scopes/{scopeID}", s.patchScope)

				o.Post("/scopes/{scopeID}/notifications", s.createChannel)
				o.Patch("/notifications/{channelID}", s.patchChannel)
				o.Delete("/notifications/{channelID}", s.deleteChannel)
				o.Post("/notifications/{channelID}/test", s.testChannel)
				o.Patch("/findings/{findingID}", s.patchFinding)

				o.Post("/wordlists", s.uploadWordlist)
				o.Patch("/wordlists/{wordlistID}", s.patchWordlist)
				o.Put("/wordlists/{wordlistID}/content", s.putWordlistContent)
				o.Delete("/wordlists/{wordlistID}", s.deleteWordlist)
			})

			// --- admin: everything that changes what the system can do ---
			r.Group(func(r chi.Router) {
				r.Use(require(auth.RoleAdmin))
				a := roled(r, "admin", s)

				a.Get("/users", s.listUsers)
				a.Post("/users", s.createUser)
				a.Patch("/users/{userID}", s.patchUser)
				a.Delete("/users/{userID}", s.deleteUser)

				// Credentials for somebody else's network.
				a.Post("/scopes/{scopeID}/vpn-configs", s.createVPNConfig)
				a.Delete("/vpn-configs/{vpnID}", s.deleteVPNConfig)

				// Enrolling a worker hands out a credential; scaling them
				// creates containers on the host.
				a.Post("/workers/enrollment-tokens", s.createEnrollmentToken)
				a.Post("/workers/{workerID}/{action}", s.workerAction)
				a.Delete("/workers/{workerID}", s.deleteWorker)
			})
		})
	})
	return r
}

// --- helpers ---

func securityHeaders(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Attacker-controlled banners/titles are rendered as text by the SPA;
		// these headers harden the app shell itself (wiki/Architecture.md §10.3).
		w.Header().Set("X-Content-Type-Options", "nosniff")
		w.Header().Set("X-Frame-Options", "DENY")
		w.Header().Set("Referrer-Policy", "no-referrer")
		w.Header().Set("Content-Security-Policy", "default-src 'self'; img-src 'self' data:; object-src 'none'; frame-ancestors 'none'")
		next.ServeHTTP(w, r)
	})
}

func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(jsonArrays(v))
}

// jsonArrays replaces nil slices with empty ones so a list endpoint with
// nothing to report answers [] rather than null.
//
// Go marshals a nil slice as null, which every store method returns when a
// query matches no rows. The SPA's types promise arrays and read .length off
// them, so an empty result crashed the page instead of rendering "none yet" —
// a scope with no findings, a run whose targets are not planned yet. Doing it
// here fixes the whole class at once; a new list handler cannot reintroduce it.
//
// Composite payloads are built as map[string]any (run activity, the asset
// graph), so those are walked too. Slices nested inside structs are not
// reachable this way and are initialized at their construction site.
func jsonArrays(v any) any {
	if m, ok := v.(map[string]any); ok {
		out := make(map[string]any, len(m))
		for k, val := range m {
			out[k] = jsonArrays(val)
		}
		return out
	}
	if rv := reflect.ValueOf(v); rv.Kind() == reflect.Slice && rv.IsNil() {
		return reflect.MakeSlice(rv.Type(), 0, 0).Interface()
	}
	return v
}

func writeErr(w http.ResponseWriter, status int, msg string) {
	writeJSON(w, status, map[string]string{"error": msg})
}

func readJSON(r *http.Request, v any) error {
	return readJSONLimit(r, v, 1<<20)
}

// readJSONLimit decodes a JSON body with an explicit size ceiling. Most
// endpoints take small documents and use readJSON's 1 MB default; the wordlist
// editor posts whole files and needs a larger one, and without this a list
// between 1 MB and the editor's own limit fails to decode and is reported as a
// missing field rather than an oversized one.
func readJSONLimit(r *http.Request, v any, max int64) error {
	defer r.Body.Close()
	return json.NewDecoder(http.MaxBytesReader(nil, r.Body, max)).Decode(v)
}

var _ = slog.Default
var _ = time.Now
var _ = domain.RunQueued

// RouteInfo is one registered route and the least role that may call it —
// recorded at registration, so the API reference and the OpenAPI document can
// be generated from the router rather than maintained beside it.
type RouteInfo struct {
	Method, Path, Role string
}

const (
	RolePublic      = "public"
	RoleAnySignedIn = "any"
)

// roleRouter registers routes exactly like chi.Router and records each one.
type roleRouter struct {
	r    chi.Router
	role string
	s    *Server
}

func roled(r chi.Router, role string, s *Server) roleRouter {
	return roleRouter{r: r, role: role, s: s}
}

func (rr roleRouter) rec(method, pattern string) {
	rr.s.routes = append(rr.s.routes, RouteInfo{Method: method, Path: "/api/v1" + pattern, Role: rr.role})
}
func (rr roleRouter) Get(p string, h http.HandlerFunc)    { rr.rec("GET", p); rr.r.Get(p, h) }
func (rr roleRouter) Post(p string, h http.HandlerFunc)   { rr.rec("POST", p); rr.r.Post(p, h) }
func (rr roleRouter) Put(p string, h http.HandlerFunc)    { rr.rec("PUT", p); rr.r.Put(p, h) }
func (rr roleRouter) Patch(p string, h http.HandlerFunc)  { rr.rec("PATCH", p); rr.r.Patch(p, h) }
func (rr roleRouter) Delete(p string, h http.HandlerFunc) { rr.rec("DELETE", p); rr.r.Delete(p, h) }

// RouteTable returns every route with its role, in registration order. It
// builds the router on a bare server, which touches no store.
func RouteTable() []RouteInfo {
	s := &Server{logins: newLoginLimiter()}
	s.Routes()
	return s.routes
}